simplify cache
