+++ src/main.rs
@@ -1,4 +1,4 @@
-use cache::layered::LayeredCache;
+use cache::Cache;
 use dotenv::dotenv;
 use std::env;
 use std::net::SocketAddr;
@@ -16,7 +16,7 @@ mod utils;
 #[derive(Clone)]
 pub struct AppData {
     data_dir: String,
-    cache: LayeredCache,
+    cache: Cache,
 }
 
 #[tokio::main]
@@ -32,15 +32,13 @@ async fn main() -> Result<(), std::io::Error> {
 
     setup_tracing::setup_tracing();
 
-    // TODO: Calculate cache size dynamically based on available memory?
-    // 1 module ~ 512Mb
-    let layered_cache = LayeredCache::try_init(2500).await?;
-
+    // 1024Mb
+    let cache = Cache::new(1024 * 1024 * 1024).await;
     let data_dir_path = env::current_dir()?.join("temp_files");
     let data_dir = data_dir_path.as_os_str().to_str().unwrap();
     let app_data = AppData {
         data_dir: String::from(data_dir),
-        cache: layered_cache,
+        cache,
     };
 
     // create data directory
