+++ src/package/process.rs
@@ -7,7 +7,7 @@ use tracing::{error, info, span, Level};
 use transform::transformer::transform_file;
 
 use crate::app_error::ServerError;
-use crate::cache::layered::LayeredCache;
+use crate::cache::Cache;
 use crate::transform;
 
 use super::package_json::PackageJSON;
@@ -291,7 +291,7 @@ pub async fn process_npm_package(
     package_name: &str,
     package_version: &str,
     data_dir: &str,
-    cache: &LayeredCache,
+    cache: &Cache,
 ) -> Result<(MinimalCachedModule, ModuleDependenciesMap), ServerError> {
     info!(
         "Started processing package: {}@{}",
@@ -351,7 +351,7 @@ pub async fn transform_module_and_cache(
     package_name: &str,
     package_version: &str,
     data_dir: &str,
-    cache: &mut LayeredCache,
+    cache: &mut Cache,
 ) -> Result<(MinimalCachedModule, ModuleDependenciesMap), ServerError> {
     let (transformed_module, module_dependencies) =
         process_npm_package(package_name, package_version, data_dir, cache).await?;
@@ -363,7 +363,7 @@ pub async fn transform_module_and_cache(
             transform_cache_key.as_str(),
             transformed_module_serialized.as_str(),
         )
-        .await?;
+        .await;
 
     let dependencies_cache_key = get_dependencies_cache_key(package_name, package_version);
     let module_dependencies_serialized = serde_json::to_string(&module_dependencies)?;
@@ -372,7 +372,7 @@ pub async fn transform_module_and_cache(
             dependencies_cache_key.as_str(),
             module_dependencies_serialized.as_str(),
         )
-        .await?;
+        .await;
 
     Ok((transformed_module, module_dependencies))
 }
@@ -380,7 +380,7 @@ pub async fn transform_module_and_cache(
 pub async fn transform_module_cached(
     package_specifier: &str,
     data_dir: &str,
-    cache: &mut LayeredCache,
+    cache: &mut Cache,
 ) -> Result<MinimalCachedModule, ServerError> {
     let (package_name, package_version) = parse_package_specifier(package_specifier)?;
 
@@ -408,7 +408,7 @@ pub async fn module_dependencies_cached(
     package_name: &str,
     package_version: &str,
     data_dir: &str,
-    cache: &mut LayeredCache,
+    cache: &mut Cache,
 ) -> Result<ModuleDependenciesMap, ServerError> {
     let transform_cache_key = get_dependencies_cache_key(package_name, package_version);
     if let Some(cached_value) = cache.get_value(transform_cache_key.as_str()).await {
